requirements.txt
PyQt5==5.15.9
PyGithub==1.59.1
gitpython==3.1.40
cryptography==41.0.7
keyring==24.1.0
requests==2.31.0